bms/mock/requirements.txt
requests==2.32.4
base58==2.1.1
pycryptodome==3.23.0